backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
python-dotenv
requests
PyYAML
python-telegram-bot==20.7
aiohttp
pytz
celery
redis